Move test to correct package
