+++ src/test/java/no/uib/inf112/desktop/HeadlessLauncher.java
